# PERF/CLN: Improve datetime-like index ops perf