@jreback 2015-06-04
yes I would add a vbench for this.

btw, I think its actually useful to use ``.hasnans`` to skip some of the masking operations


``min_stamp = self[~self._isnan].asi8.min()``
can be
```
if self._hasnans:
    self = self[~self._isnan]
min_stamp = self.asi8.min()
```
on a NO nan series that would be a big win (assume that this is not the first operation, e.g. it obviously needs caching).

here's another thing. I think you can actually do a lot of this not in the constructor but in certain operations
(this may require you to keep another attribute)

e.g.

if you construct from a ``date_range`` you KNOW that you don't have any ``NaT`` already. 
On ``insert/append`` you can see if you have any nans (in the appended one and the existing), so you don't actually have to do the check.

So it can still be lazy but in some cases you can set it a-priori w/o actually doing the check.