@sinhrks 2015-06-05
@jreback Exactly! As a first step, added ``hasnans`` conditions before mask (``_isnan``)

> I think you can actually do a lot of this not in the constructor but in certain operations
(this may require you to keep another attribute)

I assume it means adding new property to ``_attributes`` to show whether data has ``NaT`` or not. From my understanding, implementation should be:

- Adding a metadata which shows whether the data MAY contains ``NaT``, like ``_maybe_hasnans``.
- Update the value to ``False`` when we can confirm the value DOESN'T contains ``NaT``
- When ``_maybe_hasnans`` is ``False``, ``hasnans`` returns ``False`` without checking ``_isnan``.
- When ``_maybe_hasnans`` is ``True``, ``hasnans`` checks ``_isnan`` and return the result, cache itself and change ``_maybe_hasnans``.