@jreback 2015-06-05
@sinhrks yep, exactly. For instance, ``Int64Index``, will always be ``False`` for this.