@sinhrks 2015-08-31
I couldn't undertake to add an attribute which indicates `NaN`` possibility. 

Because this is PR improve a perf somewhat, can I do it separately?