@jreback 2015-10-18
I think this needs to wait for ``Index.fillna`` yes?