@jreback 2015-11-20
@sinhrks I belive we can put this in now? move whatsnew note to 0.18.0