@sinhrks 2015-12-10
@jreback Rebased, and ready for review.

I misunderstood about ``fillna``. It can't be used because the target is ``np.ndarray``.